Added options to definition

+++ Private/Clear-TempFiles.ps1
@@ -1,20 +1,21 @@
 function Clear-TempFiles {
     Param
     (
-        [Parameter(Mandatory)]
         [HyperVServer[]]$HyperVServers
     )
 
-    foreach ($server in $HyperVServers) {
-        Invoke-Command -ComputerName $server.Name {
+    if ($HyperVServers) {
+        foreach ($server in $HyperVServers) {
+            Invoke-Command -ComputerName $server.Name {
 
-            $temp = $env:TEMP
-            $tempGI = "$temp\HyperDeployGoldenImage.vhdx"
+                $temp = $env:TEMP
+                $tempGI = "$temp\HyperDeployGoldenImage.vhdx"
 
-            if (Test-Path $tempGI) {
-                Remove-Item $tempGI -Force
+                if (Test-Path $tempGI) {
+                    Remove-Item $tempGI -Force
+                }
+
             }
-
         }
     }
